requirements.txt
# Python dependencies for AI-Assisted Job Application Platform
# Install with: pip install -r requirements.txt

# Core
pydantic>=2.0.0

# AI Integration
openai>=1.0.0
google-generativeai>=0.3.0

# Resume Parsing
PyPDF2>=3.0.0
python-docx>=1.0.0
pdfplumber>=0.10.0

# Web Scraping (for job discovery)
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dateutil>=2.8.0
python-dotenv>=1.0.0

# Validation
email-validator>=2.1.0

# Image Processing (for social media)
Pillow>=10.0.0

# Cloud Storage
boto3>=1.34.0
cloudinary>=1.38.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Security
cryptography>=41.0.0
passlib>=1.7.4
